Merge latest main changes

+++ app/schemas/grocery.py
@@ -9,7 +9,7 @@ class GroceryListCreate(BaseModel):
 class GroceryListItemResponse(BaseModel):
     id: int
     ingredient: Ingredient
-    quantity: Optional[int] = None
+    quantity: Optional[float] = None
     unit: Optional[str] = None
     
     model_config = {"from_attributes": True}
@@ -25,12 +25,12 @@ class GroceryListResponse(BaseModel):
 
 class PurchaseItemInfo(BaseModel):
     itemId: int
-    quantity: int
+    quantity: float
 
 class PurchaseRequestDto(BaseModel):
     items: List[PurchaseItemInfo]
 
 class UpdateItemRequestDto(BaseModel):
-    quantity: Optional[int] = None
+    quantity: Optional[float] = None
     unit: Optional[str] = None
     expiryDate: Optional[date] = None
